+++ example/lib/example_editor/buttons/line_height_button.dart
@@ -1,23 +1,19 @@
+import 'package:example/example_editor/buttons/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_quill/extensions.dart';
 import 'package:flutter_quill/flutter_quill.dart';
 import 'package:flutter_quill/translations.dart';
-import 'package:flutter_quill_to_pdf/core/constant/constants.dart';
 import 'package:flutter_quill_to_pdf/flutter_quill_to_pdf.dart';
 
-import 'line_height_attr/line_height_attribute.dart';
-
-class QuillLineHeightButton extends QuillToolbarBaseValueButton<
-    QuillLineHeightButtonOptions, QuillLineHeightButtonExtraOptions> {
+class QuillLineHeightButton
+    extends QuillToolbarBaseValueButton<QuillLineHeightButtonOptions, QuillLineHeightButtonExtraOptions> {
   QuillLineHeightButton({
     required super.controller,
-    @Deprecated('Please use the default display text from the options')
-    this.defaultDisplayText,
+    @Deprecated('Please use the default display text from the options') this.defaultDisplayText,
     super.options = const QuillLineHeightButtonOptions(),
     super.key,
   })  : assert(options.rawItems?.isNotEmpty ?? true),
-        assert(options.initialValue == null ||
-            (options.initialValue?.isNotEmpty ?? true));
+        assert(options.initialValue == null || (options.initialValue?.isNotEmpty ?? true));
 
   final String? defaultDisplayText;
 
@@ -25,16 +21,13 @@ class QuillLineHeightButton extends QuillToolbarBaseValueButton<
   QuillLineHeightButtonState createState() => QuillLineHeightButtonState();
 }
 
-class QuillLineHeightButtonState extends QuillToolbarBaseValueButtonState<
-    QuillLineHeightButton,
-    QuillLineHeightButtonOptions,
-    QuillLineHeightButtonExtraOptions,
-    String> {
+class QuillLineHeightButtonState extends QuillToolbarBaseValueButtonState<QuillLineHeightButton,
+    QuillLineHeightButtonOptions, QuillLineHeightButtonExtraOptions, String> {
   Size? size;
   final MenuController _menuController = MenuController();
 
   List<String> get rawItemsMap {
-    const List<String> spacings = Constant.default_editor_spacing;
+    const List<String> spacings = default_editor_spacing;
     return spacings;
   }
 
@@ -47,11 +40,8 @@ class QuillLineHeightButtonState extends QuillToolbarBaseValueButtonState<
 
   @override
   String get currentStateValue {
-    final Attribute<dynamic>? attribute =
-        controller.getSelectionStyle().attributes[lineHeightKey];
-    return attribute == null
-        ? _defaultDisplayText
-        : attribute.value ?? _defaultDisplayText;
+    final Attribute<dynamic>? attribute = controller.getSelectionStyle().attributes[const LineHeightAttribute().key];
+    return attribute == null ? _defaultDisplayText : attribute.value ?? _defaultDisplayText;
   }
 
   @override
@@ -77,8 +67,8 @@ class QuillLineHeightButtonState extends QuillToolbarBaseValueButtonState<
           onPressed: () {
             final String newValue = spacing;
             final attribute0 = currentValue == spacing
-                ? const LineHeightAttribute(value: null)
-                : LineHeightAttribute(value: newValue);
+                ? const LineHeightAttribute()
+                : LineHeightAttribute(lineHeight: double.tryParse(newValue));
             controller.formatSelection(attribute0);
             setState(() {
               currentValue = newValue;
@@ -96,16 +86,11 @@ class QuillLineHeightButtonState extends QuillToolbarBaseValueButtonState<
                     children: [
                       TextSpan(
                           text: 'Spacing: ',
-                          style: TextStyle(
-                              fontWeight: currentValue.equals(spacing)
-                                  ? FontWeight.bold
-                                  : FontWeight.w300)),
+                          style:
+                              TextStyle(fontWeight: currentValue.equals(spacing) ? FontWeight.bold : FontWeight.w300)),
                       TextSpan(
                         text: spacing,
-                        style: TextStyle(
-                            fontWeight: currentValue.equals(spacing)
-                                ? FontWeight.bold
-                                : FontWeight.w300),
+                        style: TextStyle(fontWeight: currentValue.equals(spacing) ? FontWeight.bold : FontWeight.w300),
                       ),
                     ],
                   ),
@@ -168,11 +153,9 @@ class QuillLineHeightButtonState extends QuillToolbarBaseValueButtonState<
 
 /// The [T] is the options for the button
 /// The [E] is the extra options for the button
-abstract class QuillToolbarBaseValueButton<
-    T extends QuillToolbarBaseButtonOptions<T, E>,
+abstract class QuillToolbarBaseValueButton<T extends QuillToolbarBaseButtonOptions<T, E>,
     E extends QuillToolbarBaseButtonExtraOptions> extends StatefulWidget {
-  const QuillToolbarBaseValueButton(
-      {required this.controller, required this.options, super.key});
+  const QuillToolbarBaseValueButton({required this.controller, required this.options, super.key});
 
   final T options;
 
@@ -181,11 +164,8 @@ abstract class QuillToolbarBaseValueButton<
 
 /// The [W] is the widget that creates this State
 /// The [V] is the type of the currentValue
-abstract class QuillToolbarBaseValueButtonState<
-    W extends QuillToolbarBaseValueButton<T, E>,
-    T extends QuillToolbarBaseButtonOptions<T, E>,
-    E extends QuillToolbarBaseButtonExtraOptions,
-    V> extends State<W> {
+abstract class QuillToolbarBaseValueButtonState<W extends QuillToolbarBaseValueButton<T, E>,
+    T extends QuillToolbarBaseButtonOptions<T, E>, E extends QuillToolbarBaseButtonExtraOptions, V> extends State<W> {
   T get options => widget.options;
 
   QuillController get controller => widget.controller;
@@ -208,7 +188,7 @@ abstract class QuillToolbarBaseValueButtonState<
   }
 
   void _checkSelectionAttribute() {
-    final attr = controller.getSelectionStyle().attributes[lineHeightKey];
+    final attr = controller.getSelectionStyle().attributes[const LineHeightAttribute().key];
     if (attr != null) {
       // checkbox tapping causes controller.selection to go to offset 0
       // controller.formatSelection(const LineHeightAttribute(value: null));
@@ -244,9 +224,7 @@ abstract class QuillToolbarBaseValueButtonState<
   String get defaultTooltip;
 
   String get tooltip {
-    return options.tooltip ??
-        context.quillToolbarBaseButtonOptions?.tooltip ??
-        defaultTooltip;
+    return options.tooltip ?? context.quillToolbarBaseButtonOptions?.tooltip ?? defaultTooltip;
   }
 
   double get iconSize {
@@ -270,13 +248,11 @@ abstract class QuillToolbarBaseValueButtonState<
   }
 
   VoidCallback? get afterButtonPressed {
-    return options.afterButtonPressed ??
-        baseButtonExtraOptions?.afterButtonPressed;
+    return options.afterButtonPressed ?? baseButtonExtraOptions?.afterButtonPressed;
   }
 }
 
-class QuillLineHeightButtonExtraOptions
-    extends QuillToolbarBaseButtonExtraOptions {
+class QuillLineHeightButtonExtraOptions extends QuillToolbarBaseButtonExtraOptions {
   const QuillLineHeightButtonExtraOptions({
     required super.controller,
     required this.currentValue,
@@ -290,14 +266,14 @@ class QuillLineHeightButtonExtraOptions
 }
 
 @immutable
-class QuillLineHeightButtonOptions extends QuillToolbarBaseButtonOptions<
-    QuillLineHeightButtonOptions, QuillLineHeightButtonExtraOptions> {
+class QuillLineHeightButtonOptions
+    extends QuillToolbarBaseButtonOptions<QuillLineHeightButtonOptions, QuillLineHeightButtonExtraOptions> {
   const QuillLineHeightButtonOptions({
     super.iconSize,
     super.iconButtonFactor,
     this.rawItems,
     this.onSelected,
-    this.attribute = const LineHeightAttribute(value: "1.0"),
+    this.attribute = const LineHeightAttribute(lineHeight: 1.0),
     super.afterButtonPressed,
     super.tooltip,
     this.padding,