+++ lib/core/extensions/string_extension.dart
@@ -1,14 +1,7 @@
-import 'package:flutter_quill_to_pdf/core/constant/constants.dart';
-
-final StringBuffer _buffer = StringBuffer();
-
 extension StringNullableExt on String? {
   String get withBrackets {
     return '[$this]';
   }
-
-  String? replaceAllNewLinesWith(Object object) =>
-      this?.replaceAll('\n', '$object');
 
   ///Equals is a similar function that use Java or Kotlin
   ///classes to see the equality from two objects
@@ -31,12 +24,6 @@ extension StringExtension on String {
   bool get isTotallyEmpty =>
       replaceAll(RegExp(r'\s+'), '').replaceAll(RegExp('\\n|\n'), '').isEmpty;
 
-  List<String> get splitBasedNewLine => split('\n');
-
-  String replaceAllNewLinesWith(Object object) => replaceAll('\n', '$object');
-
-  String get replaceHtmlBrToManyNewLines => replaceAll('<br>', '\n');
-
   bool equals(String other,
       {bool caseSensitive = true,
       Pattern? pattern,
@@ -48,139 +35,6 @@ extension StringExtension on String {
             : pattern.allMatches(useThisInstead ? this : other).isNotEmpty
         : this == other;
   }
-
-  ///Since our default delta to html can create inline styles like
-  /// <strong style="line-height:1.0;font-size:8px">bold</strong
-  ///Instead <span style="..."><strong>bold</strong></span>
-  ///We decide create a converter that fix this minimal bug
-  String get convertWrongInlineStylesToSpans {
-    return replaceAllMapped(RegExp(r'(<em\s*(style=".*?")>(.+?)<\/em>)'),
-        (Match match) {
-      final String? styles = match.group(2);
-      final String? content = match.group(3);
-      return '<span $styles><em>$content</em></span>';
-    }).replaceAllMapped(RegExp(r'(<strong\s*(style=".*?")>(.+?)<\/strong>)'),
-        (Match match) {
-      final String? styles = match.group(2);
-      final String? content = match.group(3);
-      return '<span $styles><strong>$content</strong></span>';
-    }).replaceAllMapped(RegExp(r'(<u\s*(style=".*?")>(.+?)<\/u>)'),
-        (Match match) {
-      final String? styles = match.group(2);
-      final String? content = match.group(3);
-      return '<span $styles><u>$content</u></span>';
-    }).replaceAllMapped(RegExp(r'(<s\s*(style=".*?")>(.+?)<\/s>)'),
-        (Match match) {
-      final String? styles = match.group(2);
-      final String? content = match.group(3);
-      return '<span $styles><s>$content</s></span>';
-    });
-  }
-
-  ///Encode the markdown inline styles (**,*,_) to another tags types, to avoid detection by the compiler
-  //it's exist, because if the user put "_" in a
-  ///word, it will be detected as underline
-  ///instead a symbol, a put the underline style (the user couldn't want this)
-  //TODO: implement strike
-  String get encodeSymbols {
-    return replaceAll('#', Constant.ENCODED_MD_HEADER_SYMBOL)
-        .replaceAll('"', Constant.ENCODED_QUOTE_SYMBOL)
-        .replaceAll('}', Constant.ENCODED_KEY_SYMBOL_LEFT)
-        .replaceAll('{', Constant.ENCODED_KEY_SYMBOL_RIGHT)
-        .replaceAll(']', Constant.ENCODED_BRACKETS_SYMBOL_LEFT)
-        .replaceAll('[', Constant.ENCODED_BRACKETS_SYMBOL_RIGHT)
-        .replaceAll('***_', Constant.ENCODED_MD_ALL_SYMBOL_LEFT)
-        .replaceAll('_***', Constant.ENCODED_MD_ALL_SYMBOL_RIGHT)
-        .replaceAll('**_', Constant.ENCODED_MD_UNDERLINE_BOLD_SYMBOL_LEFT)
-        .replaceAll('_**', Constant.ENCODED_MD_UNDERLINE_BOLD_SYMBOL_RIGHT)
-        .replaceAll('*_', Constant.ENCODED_MD_UNDERLINE_ITALIC_LEFT)
-        .replaceAll('_*', Constant.ENCODED_MD_UNDERLINE_ITALIC_RIGHT)
-        .replaceAll('***', Constant.ENCODED_MD_BOLDITALIC_SYMBOL)
-        .replaceAll('_', Constant.ENCODED_MD_UNDERLINE_SYMBOL)
-        .replaceAll('**', Constant.ENCODED_MD_BOLD_SYMBOL)
-        .replaceAll('*', Constant.ENCODED_MD_ITALIC_SYMBOL);
-  }
-
-  //TODO: implement strike
-  String get decodeSymbols {
-    return replaceAll(Constant.ENCODED_MD_HEADER_SYMBOL, '#')
-        .replaceAll(Constant.ENCODED_KEY_SYMBOL_LEFT, '}')
-        .replaceAll(Constant.ENCODED_QUOTE_SYMBOL, '"')
-        .replaceAll(Constant.ENCODED_KEY_SYMBOL_RIGHT, '{')
-        .replaceAll(Constant.ENCODED_BRACKETS_SYMBOL_LEFT, ']')
-        .replaceAll(Constant.ENCODED_BRACKETS_SYMBOL_RIGHT, '[')
-        .replaceAll(Constant.ENCODED_MD_ALL_SYMBOL_LEFT, '***_')
-        .replaceAll(Constant.ENCODED_MD_ALL_SYMBOL_RIGHT, '_***')
-        .replaceAll(Constant.ENCODED_MD_UNDERLINE_BOLD_SYMBOL_LEFT, '**_')
-        .replaceAll(Constant.ENCODED_MD_UNDERLINE_BOLD_SYMBOL_RIGHT, '_**')
-        .replaceAll(Constant.ENCODED_MD_UNDERLINE_ITALIC_LEFT, '*_')
-        .replaceAll(Constant.ENCODED_MD_UNDERLINE_ITALIC_RIGHT, '_*')
-        .replaceAll(Constant.ENCODED_MD_BOLDITALIC_SYMBOL, '***')
-        .replaceAll(Constant.ENCODED_MD_UNDERLINE_SYMBOL, '_')
-        .replaceAll(Constant.ENCODED_MD_BOLD_SYMBOL, '**')
-        .replaceAll(Constant.ENCODED_MD_ITALIC_SYMBOL, '*');
-  }
-
-  ///A simple inline style html to markdown converter
-  //TODO: implement link conversion
-  String get convertHTMLToMarkdown =>
-      replaceAllMapped(RegExp(r'<strong>(?<bold>(?:(?!(<strong>|<\/strong>)).)+)<\/strong>'),
-          (Match match) {
-        return '**${match.group(1)!}**';
-      }).replaceAllMapped(RegExp(r'<em>(?<italic>(?:(?!(<em>|<\/em>)).)+)<\/em>'),
-          (Match match) {
-        return '*${match.group(1)!}*';
-      }).replaceAllMapped(RegExp(r'<s>(?<strike>(?:(?!(<s>|<\/s>)).)+)<\/s>'),
-          (Match match) {
-        return '~~${match.group(1)!}~~';
-      }).replaceAllMapped(RegExp(r'<s><u>(?<strike>(?:(?!(<s>|<\/s>)).)+)<\/u><\/s>'),
-          (Match match) {
-        return '~~_${match.group(1)!}_~~';
-      }).replaceAllMapped(RegExp(r'<em><s><u>(?<strikeItalicUnderline>(?:(?!(<em><s><u>|<\/u><\/s><\/em>)).)+)<\/u><\/s><\/em>'),
-          (Match match) {
-        return '*~~_${match.group(1)!}_~~*';
-      }).replaceAllMapped(RegExp(r'<strong><em><s><u>(?<strikeBoldItalicUnderline>(?:(?!(<strong><em><s><u>|<\/u><\/s><\/em><\/strong>)).)+)<\/u><\/s><\/em><\/strong>'),
-          (Match match) {
-        return '***~~_${match.group(1)!}_~~***';
-      }).replaceAllMapped(RegExp(r'<strong><s><u>(?<strikeBoldUnderline>(?:(?!(<strong><s><u>|<\/u><\/s><\/strong>)).)+)<\/u><\/s><\/strong>'),
-          (Match match) {
-        return '**~~_${match.group(1)!}_~~**';
-      }).replaceAllMapped(RegExp(r'<strong><s>(?<strikeBold>(?:(?!(<strong><s>|<\/s><\/strong>)).)+)<\/s><\/strong>'),
-          (Match match) {
-        return '**~~${match.group(1)!}~~**';
-      }).replaceAllMapped(RegExp(r'<u>(?<underline>(?:(?!(<u>|<\/u>)).)+)<\/u>'),
-          (Match match) {
-        return '_${match.group(1)!}_';
-      }).replaceAllMapped(RegExp(r'<em><s>(?<strikeItalic>(?:(?!(<em><s>|<\/s><\/em>)).)+)<\/s><\/em>'),
-          (Match match) {
-        return '*~~${match.group(1)!}~~*';
-      }).replaceAllMapped(RegExp(r'<em><s><u>(?<strike>(?:(?!(<em><s><u>|<\/u><\/s><\/em>)).)+)<\/u><\/s><\/em>'),
-          (Match match) {
-        return '*~~_${match.group(1)!}_~~*';
-      }).replaceAllMapped(RegExp(r'<em><u>(?<italicunderline>(?:(?!(<em><u>|<\/u><\/em>)).)+)<\/u><\/em>'),
-          (Match match) {
-        return '*_${match.group(1)!}_*';
-      }).replaceAllMapped(RegExp(r'<strong><u>(?<boldunderline>(?:(?!(<strong><u>|<\/u><\/strong>)).)+)<\/u><\/strong>'),
-          (Match match) {
-        return '**_${match.group(1)!}_**';
-      }).replaceAllMapped(RegExp(r'<strong><em>(?<italicbold>(?:(?!(<strong><em>|<\/strong><\/em>)).)+)<\/em><\/strong>'),
-          (Match match) {
-        return '***${match.group(1)!}***';
-      }).replaceAllMapped(RegExp(r'<strong><em><u>(?<bolditalicunderline>(?:(?!(<strong><em><u>|<\/u><\/em><\/strong>)).)+)<\/u><\/em><\/strong>'),
-          (Match match) {
-        return '***_${match.group(1)!}_***';
-      }).replaceAllMapped(Constant.WRONG_IMAGE_MATCHING, (Match match) {
-        final String styles = match.group(1)!;
-        final String src = match.group(12)!;
-        return '![$styles]($src)';
-      });
-
-  ///Since [vsc_quill_delta_to_html] encode the text with UTF8,
-  ///don't decode at the output, this function solve this
-  String get convertUTF8QuotesToValidString =>
-      replaceAll('&lt;', '<').replaceAll('&gt;', '>');
-  String get recovertUTF8QuotesToHumanStringChars =>
-      replaceAll('<', '&lt;').replaceAll('>', '&gt;');
 
   ///Used to solved common errors in raw delta strings, since we don't used delta literals
   String get fixCommonErrorInsertsInRawDelta => replaceAll('"}]{"', '"},{"')